# Add custom.git_status overide using spaceship-prompt implementation.