Since the default libgit2 integration in starship is so slow (for larger repos I commonly work in, its > 8s) the default `$git_status` prompt has been disabled for quite some time.

This brings back a customized version of leveraging both starship@0.45's new custom prompts (lets you run arbitrary shell commands) and [denysdovhan/spaceship-prompt](https://github.com/denysdovhan/spaceship-prompt) git status module to provide a much faster implementation (by using `git status` natively).

In smaller repos, this is a bit slower than the built in starship implementation (from 0.023s to 0.084s) but on larger repos it is **massively** faster (from 8.654s to 0.509s).